LayoutTests/editing/inserting/font-size-clears-from-typing-style.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description("Tests that we don't serialize redundant font-size in typingStyle. There should be no span or style attribute around A or B below."
     + "See https://bugs.webkit.org/show_bug.cgi?id=26279.");

var editable = document.createElement('div');
editable.contentEditable = true;
editable.innerHTML = '<br><div id="wrapper">A</div>';

document.body.appendChild(editable);
editable.focus();

window.getSelection().setBaseAndExtent(editable, 0, editable, 0);
document.execCommand('ForwardDelete', false, false);
document.execCommand('InsertText', false, 'B');

Markup.dump(editable);

</script>
</body>
</html>
