third_party/blink/web_tests/platform/generic/editing/inserting/font-size-clears-from-typing-style-expected.txt

Tests that we don't serialize redundant font-size in typingStyle. There should be no span or style attribute around A or B below.See https://bugs.webkit.org/show_bug.cgi?id=26279.
| <div>
|   id="wrapper"
|   "B<#selection-caret>A"
